ES6ify Node tests

+++ test/node/unix-sockets.js
@@ -1,33 +1,33 @@
-var request = require('../../'),
-    express = require('express'),
-    assert = require('assert'),
-    app = express(),
-    http = require('http'),
-    https = require('https'),
-    os = require('os'),
-    fs = require('fs'),
-    key = fs.readFileSync(__dirname + '/fixtures/key.pem'),
-    cert = fs.readFileSync(__dirname + '/fixtures/cert.pem'),
-    httpSockPath = [os.tmpdir(), 'superagent-http.sock'].join('/'),
-    httpsSockPath = [os.tmpdir(), 'superagent-https.sock'].join('/'),
-    httpServer,
-    httpsServer;
+"use strict";
+const request = require("../../");
+const express = require("express");
+const assert = require("assert");
+const app = express();
+const http = require("http");
+const https = require("https");
+const os = require("os");
+const fs = require("fs");
+const key = fs.readFileSync(`${__dirname}/fixtures/key.pem`);
+const cert = fs.readFileSync(`${__dirname}/fixtures/cert.pem`);
+const httpSockPath = [os.tmpdir(), "superagent-http.sock"].join("/");
+const httpsSockPath = [os.tmpdir(), "superagent-https.sock"].join("/");
+let httpServer;
+let httpsServer;
 
-if (process.platform === 'win32') {
-  return;
-}
-
-app.get('/', function(req, res) {
-  res.send('root ok!');
+app.get("/", (req, res) => {
+  res.send("root ok!");
 });
 
-app.get('/request/path', function(req, res) {
-  res.send('request path ok!');
+app.get("/request/path", (req, res) => {
+  res.send("request path ok!");
 });
 
-describe('[unix-sockets] http', function() {
+describe("[unix-sockets] http", () => {
+  if (process.platform === "win32") {
+    return;
+  }
 
-  before(function(done) {
+  before(done => {
     if (fs.existsSync(httpSockPath) === true) {
       // try unlink if sock file exists
       fs.unlinkSync(httpSockPath);
@@ -36,75 +36,72 @@ describe('[unix-sockets] http', function() {
     httpServer.listen(httpSockPath, done);
   });
 
-  var base = 'http+unix://' + httpSockPath.replace(/\//g, '%2F');
+  const base = `http+unix://${httpSockPath.replace(/\//g, "%2F")}`;
 
-  describe('request', function() {
-    it('path: / (root)', function(done) {
-      request
-        .get(base + '/')
-        .end(function(err, res) {
-          assert(res.ok);
-          assert.strictEqual('root ok!', res.text);
-          done();
-        });
+  describe("request", () => {
+    it("path: / (root)", done => {
+      request.get(`${base}/`).end((err, res) => {
+        assert(res.ok);
+        assert.strictEqual("root ok!", res.text);
+        done();
+      });
     });
 
-    it('path: /request/path', function(done) {
-      request
-        .get(base + '/request/path')
-        .end(function(err, res) {
-          assert(res.ok);
-          assert.strictEqual('request path ok!', res.text);
-          done();
-        });
+    it("path: /request/path", done => {
+      request.get(`${base}/request/path`).end((err, res) => {
+        assert(res.ok);
+        assert.strictEqual("request path ok!", res.text);
+        done();
+      });
     });
   });
 
-  after(function(done) {
+  after(done => {
     httpServer.close(done);
   });
-
 });
 
-describe('[unix-sockets] https', function() {
+describe("[unix-sockets] https", () => {
+  if (process.platform === "win32") {
+    return;
+  }
 
-  before(function(done) {
+  before(done => {
     if (fs.existsSync(httpsSockPath) === true) {
       // try unlink if sock file exists
       fs.unlinkSync(httpsSockPath);
     }
-    httpsServer = https.createServer({ key: key, cert: cert }, app);
+    httpsServer = https.createServer({ key, cert }, app);
     httpsServer.listen(httpsSockPath, done);
   });
 
-  var base = 'https+unix://' + httpsSockPath.replace(/\//g, '%2F');
+  const base = `https+unix://${httpsSockPath.replace(/\//g, "%2F")}`;
 
-  describe('request', function() {
-    it('path: / (root)', function(done) {
+  describe("request", () => {
+    it("path: / (root)", done => {
       request
-        .get(base + '/')
+        .get(`${base}/`)
         .ca(cert)
-        .end(function(err, res) {
+        .end((err, res) => {
           assert(res.ok);
-          assert.strictEqual('root ok!', res.text);
+          assert.strictEqual("root ok!", res.text);
           done();
         });
     });
 
-    it('path: /request/path', function(done) {
+    it("path: /request/path", done => {
       request
-        .get(base + '/request/path')
+        .get(`${base}/request/path`)
         .ca(cert)
-        .end(function(err, res) {
+        .end((err, res) => {
           assert(res.ok);
-          assert.strictEqual('request path ok!', res.text);
+          assert.strictEqual("request path ok!", res.text);
           done();
         });
     });
   });
 
-  after(function(done) {
+  after(done => {
     httpsServer.close(done);
   });
-
 });
